no need for base_element_collection
