--- src/easyreflectometry/sample/collections/base_element_collection.py
@@ -1,27 +1,0 @@
-from typing import List
-from typing import Optional
-
-from .base_collection import BaseCollection
-
-
-class BaseElementCollection(BaseCollection):
-    @property
-    def _dict_repr(self) -> dict:
-        """
-        A simplified dict representation.
-
-        :return: Simple dictionary
-        """
-        return {self.name: [i._dict_repr for i in self]}
-
-    def as_dict(self, skip: Optional[List[str]] = None) -> dict:
-        """
-        Create a dictionary representation of the collection.
-
-        :return: A dictionary representation of the collection
-        """
-        this_dict = super().as_dict(skip=skip)
-        this_dict['data'] = []
-        for collection_element in self:
-            this_dict['data'].append(collection_element.as_dict(skip=skip))
-        return this_dict
