+++ src/easyreflectometry/sample/collections/sample.py
@@ -138,16 +138,12 @@ class Sample(BaseCollection):
         self.subphase.thickness.enabled = False
 
     # Representation
-    def as_dict(self, skip: list = None) -> dict:
+    def as_dict(self, skip: Optional[List[str]] = None) -> dict:
         """Produces a cleaned dict using a custom as_dict method to skip necessary things.
         The resulting dict matches the parameters in __init__
 
         :param skip: List of keys to skip, defaults to `None`.
         """
-        if skip is None:
-            skip = []
         this_dict = super().as_dict(skip=skip)
-        for i, assembly in enumerate(self.data):
-            this_dict['data'][i] = assembly.as_dict(skip=skip)
         this_dict['populate_if_none'] = self.populate_if_none
         return this_dict
